requirements.txt
openai
pyautogen==0.2.6
lumibot
alpaca-trade-api
python-timedelta
python-dotenv
pydantic<2.0
